Insert CSRF protection middleware
Added a CSRF protection middleware that checks to see if the origin of the request is the same as the host header. If it is not, then force re-auth by redirecting to the login page. This action was chosen to deal with CSRF because codebase performs similar actions when security checks are violated.